backend/requirements.txt
fastapi
uvicorn
sqlalchemy
passlib[bcrypt]
bcrypt>=4.0.1,<4.1
python-jose[cryptography]